requirements.txt
django
django-extensions
django-widget-tweaks
python-slugify